+++ sltl/sltl.vcxproj.filters
@@ -34,6 +34,24 @@
     <Filter Include="Header Files\expression">
       <UniqueIdentifier>{61dcc72a-b82c-4b0b-a9f2-cad8f2d638c4}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Header Files\output">
+      <UniqueIdentifier>{b19ff030-c137-471b-a360-56d0e7f8adbb}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Header Files\output\glsl">
+      <UniqueIdentifier>{bc55f2a8-fb1f-4a80-8acb-e468ea9868c2}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Header Files\output\hlsl">
+      <UniqueIdentifier>{ca070abe-e82f-4170-9c1b-6e584ee74f91}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files\output">
+      <UniqueIdentifier>{9bfaf5c3-9a28-4c00-a5f8-c58851835bac}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files\output\glsl">
+      <UniqueIdentifier>{b77c3e6a-2c95-4ba8-ace0-13aa69994ef5}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files\output\hlsl">
+      <UniqueIdentifier>{c589984d-d2a2-47d4-a6fd-fdd5241b19a3}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\scope.h">
@@ -65,9 +83,6 @@
     </ClInclude>
     <ClInclude Include="src\syntax\variable_declaration.h">
       <Filter>Header Files\syntax</Filter>
-    </ClInclude>
-    <ClInclude Include="src\output.h">
-      <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="src\syntax\operator.h">
       <Filter>Header Files\syntax</Filter>
@@ -246,6 +261,24 @@
     <ClInclude Include="src\detail\algorithm.h">
       <Filter>Header Files\detail</Filter>
     </ClInclude>
+    <ClInclude Include="src\output\output.h">
+      <Filter>Header Files\output</Filter>
+    </ClInclude>
+    <ClInclude Include="src\output\glsl\output_glsl.h">
+      <Filter>Header Files\output\glsl</Filter>
+    </ClInclude>
+    <ClInclude Include="src\output\output_introspector.h">
+      <Filter>Header Files\output</Filter>
+    </ClInclude>
+    <ClInclude Include="src\output\output_matrix_order.h">
+      <Filter>Header Files\output</Filter>
+    </ClInclude>
+    <ClInclude Include="src\output\glsl\glsl_language.h">
+      <Filter>Header Files\output\glsl</Filter>
+    </ClInclude>
+    <ClInclude Include="src\output\glsl\glsl_convention.h">
+      <Filter>Header Files\output\glsl</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\scope.cpp">
@@ -259,9 +292,6 @@
     </ClCompile>
     <ClCompile Include="src\syntax\block_manager.cpp">
       <Filter>Source Files\syntax</Filter>
-    </ClCompile>
-    <ClCompile Include="src\output.cpp">
-      <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="src\language.cpp">
       <Filter>Source Files</Filter>
@@ -303,6 +333,27 @@
       <Filter>Source Files\syntax</Filter>
     </ClCompile>
     <ClCompile Include="src\syntax\parameter_declaration.cpp">
+      <Filter>Source Files\syntax</Filter>
+    </ClCompile>
+    <ClCompile Include="src\output\output.cpp">
+      <Filter>Source Files\output</Filter>
+    </ClCompile>
+    <ClCompile Include="src\output\glsl\output_glsl.cpp">
+      <Filter>Source Files\output\glsl</Filter>
+    </ClCompile>
+    <ClCompile Include="src\output\output_introspector.cpp">
+      <Filter>Source Files\output</Filter>
+    </ClCompile>
+    <ClCompile Include="src\output\output_matrix_order.cpp">
+      <Filter>Source Files\output</Filter>
+    </ClCompile>
+    <ClCompile Include="src\output\glsl\glsl_language.cpp">
+      <Filter>Source Files\output\glsl</Filter>
+    </ClCompile>
+    <ClCompile Include="src\output\glsl\glsl_convention.cpp">
+      <Filter>Source Files\output\glsl</Filter>
+    </ClCompile>
+    <ClCompile Include="src\syntax\intrinsic_declaration.cpp">
       <Filter>Source Files\syntax</Filter>
     </ClCompile>
   </ItemGroup>
