+++ sltl/src/syntax/intrinsic_declaration.h
@@ -11,13 +11,15 @@ namespace sltl
 {
 namespace syntax
 {
+  const wchar_t* to_intrinsic_string(core::intrinsic id);
+
   class intrinsic_declaration : public declaration, public node
   {
   public:
     typedef std::unique_ptr<intrinsic_declaration> ptr;
 
     intrinsic_declaration(const intrinsic_declaration&) = delete;
-    intrinsic_declaration(core::intrinsic i, parameter_list&& parameters, const language::type& type_return) : declaration(language::to_intrinsic_string(i)), _intrinsic(i), _type_return(type_return), _parameters(std::move(parameters)) {}
+    intrinsic_declaration(core::intrinsic i, parameter_list&& parameters, const language::type& type_return) : declaration(to_intrinsic_string(i)), _intrinsic(i), _type_return(type_return), _parameters(std::move(parameters)) {}
 
     bool apply_action(action& act) override
     {
