+++ sltl_test/src/call_test.cpp
@@ -2,10 +2,11 @@
 
 #include "call.h"
 #include "shader.h"
-#include "output.h"
 #include "scalar.h"
 #include "vector.h"
 #include "basic_operators.h"
+
+#include "output/output.h"
 
 #include "io/io.h"
