+++ sltl_test/src/matrix_test.cpp
@@ -3,8 +3,10 @@
 #include "matrix.h"
 #include "basic_operators.h"
 #include "shader.h"
-#include "output.h"
 #include "scope.h"
+
+#include "output/output.h"
+#include "output/output_matrix_order.h"
 
 
 namespace
