+++ sltl_test/src/shader_test.cpp
@@ -1,7 +1,8 @@
 #include <gtest/gtest.h>
 
 #include "shader.h"
-#include "output.h"
+
+#include "output/output.h"
 
 #include "io/io.h"
