+++ sltl_test/src/vector_test.cpp
@@ -3,7 +3,8 @@
 #include "vector.h"
 #include "basic_operators.h"
 #include "shader.h"
-#include "output.h"
+
+#include "output/output.h"
 
 
 namespace
